feat: decompose ChatView.swift — extract markdown parser, timestamp divider, fix design tokens
